Refactor-part 1, tests passing, lint complains about undef functions
